+++ sdks/go/test/integration/io/fhirio/fhirio_test.go
@@ -24,6 +24,7 @@ import (
 	"flag"
 	"fmt"
 	"math/big"
+	"net/http"
 	"os"
 	"strconv"
 	"strings"
@@ -62,11 +63,20 @@ func checkFlags(t *testing.T) {
 	}
 }
 
+func setupFhirStoreWithData(t *testing.T) (string, []string, func()) {
+	return setupFhirStore(t, true)
+}
+
+func setupEmptyFhirStore(t *testing.T) (string, func()) {
+	storePath, _, teardownFunc := setupFhirStore(t, false)
+	return storePath, teardownFunc
+}
+
 // Sets up a test fhir store by creating and populating data to it for testing
 // purposes. It returns the name of the created store path, a slice of the
 // resource paths to be used in tests, and a function to teardown what has been
 // set up.
-func setupFhirStore(t *testing.T) (string, []string, func()) {
+func setupFhirStore(t *testing.T, shouldPopulateStore bool) (string, []string, func()) {
 	t.Helper()
 	if storeService == nil || storeManagementService == nil {
 		t.Fatal("Healthcare Services were not initialized")
@@ -75,13 +85,16 @@ func setupFhirStore(t *testing.T) (string, []string, func()) {
 	healthcareDataset := fmt.Sprintf(datasetPathFmt, *gcpopts.Project, *gcpopts.Region)
 	createdFhirStore, err := createStore(healthcareDataset)
 	if err != nil {
-		t.Fatal("Test store failed to be created")
+		t.Fatalf("Test store failed to be created. Reason: %v", err.Error())
 	}
 	createdFhirStorePath := createdFhirStore.Name
 
-	resourcePaths := populateStore(createdFhirStorePath)
-	if len(resourcePaths) == 0 {
-		t.Fatal("No data got populated to test")
+	var resourcePaths []string
+	if shouldPopulateStore {
+		resourcePaths = populateStore(createdFhirStorePath)
+		if len(resourcePaths) == 0 {
+			t.Fatal("No data got populated to test")
+		}
 	}
 
 	return createdFhirStorePath, resourcePaths, func() {
@@ -168,7 +181,7 @@ func TestFhirIO_Read(t *testing.T) {
 	integration.CheckFilters(t)
 	checkFlags(t)
 
-	_, testResourcePaths, teardownFhirStore := setupFhirStore(t)
+	_, testResourcePaths, teardownFhirStore := setupFhirStoreWithData(t)
 	defer teardownFhirStore()
 
 	p, s, resourcePaths := ptest.CreateList(testResourcePaths)
@@ -183,7 +196,7 @@ func TestFhirIO_InvalidRead(t *testing.T) {
 	integration.CheckFilters(t)
 	checkFlags(t)
 
-	fhirStorePath, _, teardownFhirStore := setupFhirStore(t)
+	fhirStorePath, _, teardownFhirStore := setupFhirStoreWithData(t)
 	defer teardownFhirStore()
 
 	invalidResourcePath := fhirStorePath + "/fhir/Patient/invalid"
@@ -192,9 +205,26 @@ func TestFhirIO_InvalidRead(t *testing.T) {
 	passert.Count(s, failedReads, "", 1)
 	passert.Empty(s, resources)
 	passert.True(s, failedReads, func(errorMsg string) bool {
-		return strings.Contains(errorMsg, "bad status [404]")
+		return strings.Contains(errorMsg, strconv.Itoa(http.StatusNotFound))
 	})
 
+	ptest.RunAndValidate(t, p)
+}
+
+func TestFhirIO_ExecuteBundles(t *testing.T) {
+	integration.CheckFilters(t)
+	checkFlags(t)
+
+	fhirStorePath, teardownFhirStore := setupEmptyFhirStore(t)
+	defer teardownFhirStore()
+
+	p, s, bundles := ptest.CreateList(readPrettyBundles())
+	successBodies, failures := fhirio.ExecuteBundles(s, fhirStorePath, bundles)
+	passert.Count(s, successBodies, "", 2)
+	passert.Count(s, failures, "", 2)
+	passert.True(s, failures, func(errorMsg string) bool {
+		return strings.Contains(errorMsg, strconv.Itoa(http.StatusBadRequest))
+	})
 	ptest.RunAndValidate(t, p)
 }